+++ R/netmeta-internal.R
@@ -25,3 +25,27 @@ invmat <- function(X) {
   ##
   res
 }
+
+
+calcV <- function(x, sm) {
+  p2 <- (x$event2[1] + x$incr[1]) / (x$n2[1] + 2 * x$incr[1])
+  n2 <- x$n2[1] + 2 * x$incr[1]
+  ##
+  if (sm == "OR")
+    V <- matrix(1 / (x$event2[1] + x$incr) +
+                1 / (x$n2[1] - x$event2[1] + x$incr[1]),
+                nrow = nrow(x), ncol = nrow(x))
+  else if (sm == "RR")
+    V <- matrix((1 - p2) / (n2 * p2),
+                nrow = nrow(x), ncol = nrow(x))
+  else if (sm == "RD")
+    V <- matrix(p2 * (1 - p2) / n2,
+                nrow = nrow(x), ncol = nrow(x))
+  else if (sm == "ASD")
+    V <- matrix(0.25 * (1 / x$n2[1]),
+                nrow = nrow(x), ncol = nrow(x))
+  ##
+  diag(V) <- x$seTE^2
+  ##
+  V
+}